LibWeb: Split BackgroundSizeStyleValue out of StyleValue.{h,cpp}
